Adds `-l <argument>` command which would list all the <argument>
Removes `--themelist` command as it is included within the `-l` command. As of now this list command, can list all the themes, languages, boxes support by piston-cli.

This commit also adds another dependency `more_itertools` which helps us group the list into pairs of two.

A basic idea used is to format the parameters in the list command in a tuple, if this list command doesn't use the default make table then don't make it a tuple and rather call the function itself. While fetching it would check whether this is a tuple or not, and if it is a tuple, then it would use the make table function else it would just call whatever is returned by dict